Moving DocumentPopupPicker to modules/Documents
